+++ examples/with-rax-pha/src/CustomTabBar/index.tsx
@@ -1,0 +1,15 @@
+import { createElement, PureComponent } from 'rax';
+import View from 'rax-view';
+import Text from 'rax-text';
+
+class CustomTabBar extends PureComponent {
+  render() {
+    return (
+      <View>
+        <Text>这是自定义 TabBar</Text>
+      </View>
+    );
+  }
+}
+
+export default CustomTabBar;