+++ packages/plugin-rax-web/src/index.ts
@@ -36,14 +36,8 @@ export default (api) => {
     validation: 'object',
   });
 
-  // Set Entry
-  setEntry(chainConfig, context);
-  registerTask(target, chainConfig);
-  // Set global methods
-  setRegisterMethod(api);
-
   if (webConfig.pha) {
-    // Modify mpa config
+    // Modify mpa config must before than set entry, because entry check depends on mpa config
     modifyUserConfig(() => {
       if (!context.userConfig.web) context.userConfig.web = {};
       context.userConfig.web.mpa = true;
@@ -51,6 +45,12 @@ export default (api) => {
       return context.userConfig;
     });
   }
+
+  // Set Entry
+  setEntry(chainConfig, context);
+  registerTask(target, chainConfig);
+  // Set global methods
+  setRegisterMethod(api);
 
   if (documentPath) {
     setLocalBuilder(api, documentPath);